fix: 디렉토리 이름 변경 및 파일 확장자 변경 (#67)
