+++ Client/src/component/Commons/NavBar.tsx
