+++ Client/src/component/Main/styled.ts
