+++ Client/src/component/Myinfo/S3ImgInput.tsx
